DEV: Split workflows expression editor CodeMirror extension into modules
The expression editor kept its whole CodeMirror integration in one ~970-line
file, mixing about ten unrelated extensions (completions, hover tooltips,
signature help, drag-and-drop, validation, decorations, theme and the
server-side evaluator) along with their helpers.

Move each extension into its own file under an expression-extensions/ directory,
with index.js assembling the extension list. This is the one-file-per-extension
layout already used under prosemirror/extensions, and the name pairs with the
sibling json-editor-extensions bundle.

The decorations StateField and the evaluator ViewPlugin only communicate through
a private effect, so they are combined into expression-evaluation.js to keep that
effect file-local. No behavior changes; the builder is renamed
buildExpressionExtensions and its one caller updated.
